Allow {name: value} for CHOICE data.
Previously only (name, value) was allowed. With this small change,
parsing formats that are not aware of the ASN.1 schema becomes trivial.

For instance we can now parse a JSON5 file and feed it to the encoder.
That means that people writing files by hand can use JSON5, which is
much more pleasant for a human than JER or (arguably) XER.  The output
of asn1vnparser can also be used, so people can use a subset of the
ASN.1 value notation.

There are probably still schemas where parsing without knowledge of the
schema doesn't work, but this catches a very common case where the tools
now become easier to use.